backend/requirements.txt
fastapi
uvicorn
tika
pytesseract
Pillow
PyYAML
sqlalchemy
psycopg2-binary
cryptography
setuptools<81
